requirements.txt
Keras~=2.1.5
matplotlib
PyFunctional
numpy
tensorboard~=1.*
tensorflow~=1.*
IPython
pydot~=1.2.4
pydot-ng~=1.0.0
setuptools
graphviz
jupyter
